+++ src/main/kotlin/uk/co/nickthecoder/paratask/project/table/Column.kt
@@ -8,7 +8,7 @@ import uk.co.nickthecoder.paratask.util.uncamel
 
 open class Column<R, T>(
         val name: String,
-        width: Double? = null,
+        width: Int? = null,
         override val label: String = name.uncamel(),
         val getter: (R) -> T
 
@@ -19,7 +19,7 @@ open class Column<R, T>(
         setCellValueFactory { p -> p.getValue().observable(name, getter) as ObservableValue<T> }
         setEditable(false)
         if (width != null) {
-            prefWidth = width
+            prefWidth = width.toDouble()
         }
     }
 }
